requirements.txt
aiogram==2.25.1
python-dotenv
aiofiles
aiohttp==3.8.6